+++ packages/web-frontend/src/actions.js
@@ -141,6 +141,7 @@ export const REQUEST_PROJECT_ACCESS = 'REQUEST_PROJECT_ACCESS';
 export const UPDATE_HISTORY_LOCATION = 'UPDATE_HISTORY_LOCATION';
 export const UPDATE_MEASURE_DATE_RANGE_ONCE_HIERARCHY_LOADS =
   'UPDATE_MEASURE_DATE_RANGE_ONCE_HIERARCHY_LOADS';
+export const HANDLE_INVALID_PERMISSION = 'HANDLE_INVALID_PERMISSION';
 
 /**
  * Attempt password change using old password, new password and new password
