+++ packages/web-frontend/src/historyNavigation/historyMiddleware.js
@@ -29,14 +29,11 @@ import {
   updateCurrentMeasureConfigOnceHierarchyLoads,
   updateHistoryLocation,
   UPDATE_MEASURE_CONFIG,
+  HANDLE_INVALID_PERMISSION,
 } from '../actions';
 import { DEFAULT_PROJECT_CODE } from '../defaults';
-import { setProject, setRequestingAccess } from '../projects/actions';
-import {
-  selectCurrentPeriodGranularity,
-  selectMeasureBarItemById,
-  selectProjectByCode,
-} from '../selectors';
+import { setProject } from '../projects/actions';
+import { selectCurrentPeriodGranularity, selectMeasureBarItemById } from '../selectors';
 import { PASSWORD_RESET_PREFIX, URL_COMPONENTS, VERIFY_EMAIL_PREFIX } from './constants';
 import {
   addPopStateListener,
@@ -48,34 +45,11 @@ import {
   PROJECTS_WITH_LANDING_PAGES,
   PROJECT_LANDING,
   LANDING,
-  REQUEST_PROJECT_ACCESS,
 } from '../containers/OverlayDiv/constants';
 import { convertDateRangeToUrlPeriodString, decodeLocation } from './utils';
 
 const userHasAccess = (state, currentProject) =>
   state.project.projects.filter(p => p.hasAccess).find(p => p.code === currentProject);
-
-// Todo: move this to a saga
-const handleInvalidPermission = async (state, dispatch, projectCode) => {
-  const { isUserLoggedIn } = state.authentication;
-
-  if (isUserLoggedIn) {
-    // show project access dialog
-    const project = selectProjectByCode(state, projectCode);
-
-    if ('code' in project) {
-      await dispatch(setRequestingAccess(project));
-      await dispatch(setOverlayComponent(REQUEST_PROJECT_ACCESS));
-      return;
-    }
-
-    // handle 404s
-    console.log('project does not exist - 404'); // Todo: handle 404s
-    return;
-  }
-  // show login dialog
-  await dispatch(setOverlayComponent(LANDING));
-};
 
 export const reactToLocationChange = (store, location, previousLocation) => {
   const { dispatch: rawDispatch } = store;
@@ -87,7 +61,7 @@ export const reactToLocationChange = (store, location, previousLocation) => {
   const hasAccess = userHasAccess(state, otherComponents.PROJECT);
 
   if (!projectSelector && !hasAccess) {
-    handleInvalidPermission(state, dispatch, otherComponents.PROJECT);
+    dispatch({ type: HANDLE_INVALID_PERMISSION, projectCode: otherComponents.PROJECT });
     return;
   }
